pyproject.toml
[tool.poetry]
name = "zem-wechat-mini-program"
version = "0.1.0"
description = "微信小程序后端服务，提供登录和柜号查询功能"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
pytz = "^2024.1"
pydantic = "^2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
